+++ client_code/lendor_registration_form/lender_registration_form_2/lender_registration_Institutional_form_2/__init__.py
@@ -9,84 +9,66 @@ import anvil.tables.query as q
 from anvil.tables import app_tables
 from datetime import datetime
 
+
 class lender_registration_Institutional_form_2(lender_registration_Institutional_form_2Template):
-  def __init__(self, user_id,**properties):
+  def __init__(self,user_id, **properties):
     self.userId = user_id
     # Set Form properties and Data Bindings.
     self.init_components(**properties)
-    user_data=app_tables.fin_user_profile.get(customer_id=user_id)
+    user_data = anvil.server.call('get_user_data', user_id)
+        
     if user_data:
-      self.text_box_1.text=user_data['business_name']
-      self.text_box_2.text=user_data['business_add']
-      self.drop_down_1.selected_value=user_data['business_type']
-      # options_2 = app_tables.fin_borrower_no_of_employees.search()
-      # option_strings_2 = [str(option['borrower_no_of_employees']) for option in options_2]
-      # self.drop_down_2.items = option_strings_2
-      self.drop_down_2.selected_value = user_data['employees_working']
-      
-      user_data.update()
-    # user_data = anvil.server.call('get_user_data', user_id)
-        
+            self.industry_type = user_data.get('industry_type', '')
+            self.turn_over = user_data.get('six_month_turnover', '')
+            self.year = user_data.get('year_estd', '')
+    
+    else:
+        self.industry_type = ''
+        self.turn_over = ''
+        self.year = ''
+
+    # user_data=app_tables.fin_user_profile.get(customer_id=user_id)
     # if user_data:
-    #         self.business_type = user_data.get('business_type', '')
-    #         self.empolyees_working = user_data.get('employees_working', '')
-    #         # self.year = user_data.get('year_estd', '')
-    #         self.business_name = user_data.get('business_name', '')
-    #         self.business_add = user_data.get('business_add', '')
-  
-            
-    # else:
-    #     self.business_type = ''
-    #     self.empolyees_working = ''
-    #     # self.year = ''
-    #     self.business_name = ''
-    #     self.business_add = ''
-        
-
+    #   self.date_picker_1.date = user_data['year_estd']
+    #   self.text_box_1.text=user_data['industry_type']
+    #   self.text_box_2.text=user_data['six_month_turnover']
     #    #Restore previously entered data if available
-    # if self.business_type:
-    #         self.drop_down_1.items = self.business_type
-    # if self.empolyees_working:
-    #        self.drop_down_2.items = self.empolyees_working
-    # # if self.year:
-    # #        self.date_picker_1.date = self.year
+    if self.industry_type:
+            self.text_box_1.text= self.industry_type
+    if self.turn_over:
+            self.text_box_2.text= self.turn_over
 
-    # if self.business_name:
-    #         self.text_box_1.text= self.business_name
-    # if self.business_add:
-    #         self.text_box_2.text= self.business_add
-
-    options = app_tables.fin_lendor_business_type.search()
-    options_string = [str(option['lendor_business_type']) for option in options]
-    self.drop_down_1.items = options_string
-
-    options = app_tables.fin_lendor_no_of_employees.search()
-    options_string = [str(option['lendor_no_of_employees']) for option in options]
-    self.drop_down_2.items = options_string
-
+    if self.year:
+           self.date_picker_1.date = self.year
+    
+    # Any code you write here will run before the form opens.
 
   def button_2_click(self, **event_args):
-    business_type = self.drop_down_1.selected_value
-    empolyees_working = self.drop_down_2.selected_value
-    # year = self.date_picker_1.date
-    business_name = self.text_box_1.text
-    business_add = self.text_box_2.text
+    industry_type = self.text_box_1.text
+    turn_over = self.text_box_2.text
+    year = self.date_picker_1.date
+    last_six_statements = self.file_loader_1.file
     user_id = self.userId
-    if not business_type or not empolyees_working or not business_add or not business_add:
+    if not industry_type or not turn_over or not last_six_statements or not year:
       Notification("Please fill all the fields").show()
     else:
-     # today = datetime.today()
-     # months = today.year * 12 + today.month - year.year * 12 - year.month
-     anvil.server.call('add_lendor_individual_form_2',business_type,empolyees_working,business_name,business_add,user_id)
+     today = datetime.today()
+     months = today.year * 12 + today.month - year.year * 12 - year.month
+     anvil.server.call('add_lendor_institutional_form_3',year,months,industry_type,turn_over,last_six_statements,user_id)
      open_form('lendor_registration_form.lender_registration_form_2.lender_registration_Institutional_form_3',user_id = user_id)
     """This method is called when the button is clicked"""
 
   def button_1_click(self, **event_args):
     user_id = self.userId
-    open_form('lendor_registration_form.lender_registration_form_2',user_id = user_id)
+    open_form('lendor_registration_form.lender_registration_form_2.lender_registration_Institutional_form_1',user_id = user_id)
     """This method is called when the button is clicked"""
 
   def button_3_click(self, **event_args):
     """This method is called when the button is clicked"""
     open_form("bank_users.user_form")
+
+  def file_loader_1_change(self, file, **event_args):
+    """This method is called when a new file is loaded into this FileLoader"""
+    if file:
+      self.image_1.source = self.file_loader_1.file
